pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepguard-workspace"
version = "0.1.0"
description = "DeepGuard (玄武) Compliance Engine — monorepo workspace"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    "pydantic>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "redis>=5.0.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "fpdf2>=2.7.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.0",
    "httpx>=0.28.0",
    "fakeredis[lua]>=2.23.0",
    "types-PyYAML>=6.0.0",
    "moto[s3]>=5.0.0",
    "pyright>=1.1.390",
    "pre-commit>=4.0.0",
    "import-linter>=2.0",
    "deptry>=0.21.0",
    "pip-audit>=2.7.0",
    "bandit[toml]>=1.8.0",
]
docs = [
    "mkdocs-material>=9.5.0",
]

[tool.setuptools.packages.find]
where = [
    "packages/core/src",
    "packages/graph/src",
    "packages/policies/src",
    "packages/parsers/src",
    "packages/connectors/src",
    "packages/agents/src",
    "packages/reporting/src",
    "packages/observability/src",
    "apps/api/src",
    "apps/worker/src",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = [
    "packages",
    "apps",
    "tests",
    "packages/core/tests",
    "packages/graph/tests",
    "packages/policies/tests",
    "packages/parsers/tests",
    "packages/connectors/tests",
    "packages/agents/tests",
    "packages/reporting/tests",
    "packages/observability/tests",
    "apps/api/tests",
    "eval/harness/tests",
]
# Legacy repo-root doc/diagram generators — not part of the Python monorepo (L0+).
extend-exclude = [
    "diagrams.py",
    "diagrams_cn.py",
    "generate_executive_summary.py",
    "generate_pdf.py",
    "generate_pdf_cn.py",
    # Large doc generator; keep linted separately if needed.
    "scripts/generate_ac_details_and_squad_docs.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"packages/**/tests/**" = ["S101"]
"eval/harness/tests/**" = ["S101"]
"packages/core/tests/**" = ["S101"]
"alembic/versions/**" = ["E501", "I001"]
"tests/integration/**" = ["E501", "SIM117"]
"scripts/seed_dev_tenant.py" = ["E501"]
"scripts/validate_user_stories_traceability.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
# So `mypy -p deepguard_*` resolves without relying on implicit namespace heuristics.
mypy_path = "packages/core/src:packages/graph/src:packages/policies/src:packages/parsers/src:packages/connectors/src:packages/agents/src:packages/reporting/src:packages/observability/src:apps/api/src:apps/worker/src"
packages = [
    "deepguard_core",
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
    "deepguard_api",
    "deepguard_worker",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
include = [
    "packages",
    "apps",
    "tests",
    "eval",
    "scripts",
    "alembic",
]
exclude = [
    ".venv",
    "**/__pycache__",
    "diagrams.py",
    "diagrams_cn.py",
    "generate_executive_summary.py",
    "generate_pdf.py",
    "generate_pdf_cn.py",
    "scripts/generate_ac_details_and_squad_docs.py",
]
extraPaths = [
    "packages/core/src",
    "packages/graph/src",
    "packages/policies/src",
    "packages/parsers/src",
    "packages/connectors/src",
    "packages/agents/src",
    "packages/reporting/src",
    "packages/observability/src",
    "apps/api/src",
    "apps/worker/src",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "tests/e2e",
    "eval/harness/tests",
    "packages/core/tests",
    "packages/graph/tests",
    "packages/policies/tests",
    "packages/parsers/tests",
    "packages/connectors/tests",
    "packages/agents/tests",
    "packages/reporting/tests",
    "packages/observability/tests",
    "apps/api/tests",
]
pythonpath = ["."]
addopts = "-q --strict-markers -m \"not integration\""
markers = [
    "req: requirement traceability AC-DG-* (value is AC id)",
    "unit: fast unit tests",
    "integration: requires Docker stack",
]

[tool.coverage.run]
branch = true
source = [
    "deepguard_core",
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
    "deepguard_api",
    "deepguard_worker",
]
omit = [
    "*/tests/*",
    # Infinite loop; exercised manually with compose. Ingest + stream helpers are unit-tested.
    "*/deepguard_worker/run.py",
    "*/deepguard_worker/__main__.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true

# deptry: editable monorepo installs all ``deepguard_*`` via one distribution; DEP003 is a false
# positive for first-party imports. DEP002 for runtime/dev tools is handled via per_rule_ignores.
[tool.deptry]
package_module_name_map = { pyright = "pyright" }
known_first_party = [
    "deepguard_core",
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
    "deepguard_api",
    "deepguard_worker",
]
extend_exclude = [
    "diagrams\\.py",
    "diagrams_cn\\.py",
    "generate_executive_summary\\.py",
    "generate_pdf\\.py",
    "generate_pdf_cn\\.py",
    "scripts/generate_ac_details_and_squad_docs\\.py",
]

[tool.deptry.per_rule_ignores]
DEP002 = [
    "asyncpg",
    "uvicorn",
    "ruff",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-html",
    "httpx",
    "fakeredis",
    "types-PyYAML",
    "pyright",
    "alembic",
    "sqlalchemy",
    "psycopg",
    "boto3",
    "moto",
    "pre-commit",
    "import-linter",
    "deptry",
    "pip-audit",
    "bandit",
]
DEP003 = [
    "starlette",
    "botocore",
    "langchain_core",
]

[tool.importlinter]
root_packages = [
    "deepguard_core",
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
    "deepguard_api",
    "deepguard_worker",
]

[[tool.importlinter.contracts]]
name = "deepguard_core is independent of other deepguard packages"
type = "forbidden"
source_modules = ["deepguard_core"]
forbidden_modules = [
    "deepguard_api",
    "deepguard_worker",
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
]

[[tool.importlinter.contracts]]
name = "Library packages do not depend on apps"
type = "forbidden"
source_modules = [
    "deepguard_graph",
    "deepguard_policies",
    "deepguard_parsers",
    "deepguard_connectors",
    "deepguard_agents",
    "deepguard_reporting",
    "deepguard_observability",
]
forbidden_modules = ["deepguard_api", "deepguard_worker"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]
